+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Imported Signal from Systemic
 * Migrated to use effigy in CI workflow
 * Fixed PHP8.1 testing
 * Updated composer check script